// src/app/modules/business-common/src/lib/components/search-summary/enums/translation-keys.enum.ts
export enum TranslationKeys {
  City_KeyNode = 'City.',
  City_BOG = 'City.BOG',
  City_MDE = 'City.MDE',
  PassengerTypes_KeyNode = 'PassengerTypes.',
  PassengerTypes_ADT = 'PassengerTypes.ADT',
  PassengerTypes_CHD = 'PassengerTypes.CHD',
  PassengerTypes_INF = 'PassengerTypes.INF',
  PassengerTypes_Plural_KeyNode = 'PassengerTypes.Plural_',
  PassengerTypes_Plural_ADT = 'PassengerTypes.Plural_ADT',
  PassengerTypes_Plural_CHD = 'PassengerTypes.Plural_CHD',
  PassengerTypes_Plural_INF = 'PassengerTypes.Plural_INF',
}